WIP: Add a (non-working) jetbrains-toolbox pacscript
The JetBrains Toolbox is a beast! It does not conform to freedesktop
standards, installs ALWAYS into ~/.local/share/JetBrains/Toolbox and
keeps its state and configuration in the same directory, making it
impossible (or impractically difficult) to create a system-wide
installable package. To make things worse, the binary distributed is
actually a self-extracting app image (yes, there is such an anymal).

There are different approaches to install it via `pacsctall`:

1. Create a chroot and have it install itself there - too complicated
   and slow; left out.

2. Make ~/.local/share/JetBrains to be a symlink to /opt/JetBrains and
   this way the installation will end up in /opt/JetBrains/Toolbox -
   this seems nice but does not work either (reason unknown).

This commit contains an attempt to implement the second approach. Kept
for historical reasons, as it doesn't work well and has been left out
as with approach 1. (There are also other ways, to appear in subsequent
commits.)