WIP: Develop further jetbrains-toolbox pacscript
The scrips now installs the Toolbox into it's default location, kills
the jetbrains-toolbox process, moves and packages the installation and
runs the application again at post-install time to allow for finalizing
the configuration.

Seems to work well. Needs code cleanup and providing a custom settings
file. (And more testing, of course ;)